feat: autenticação na rota, instalação do nodemon

+++ backend/app.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const mongoose = require('mongoose');
+const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
 
 const Usuario = require('./models/usuario');
 const UsuarioResponsavel = require('./models/usuarioResponsavel')
@@ -10,6 +12,7 @@ const Figura  = require('./models/figura')
 
 
 const mongoURI = 'mongodb+srv://saulopontes:sauloPontesBlue@clusterbluetalks.kd8yiz7.mongodb.net/blueTalksDB';
+const jwtSECRET = "bluetalks";
 
 const options = {
     useNewUrlParser: true,
@@ -38,34 +41,43 @@ const options = {
 
     //API Usuario
     
-    router.get('/usuarios/:id', (req, res) => {
-    const id = req.params.id;
-    Usuario.findById(id)
-      .then((usuario) => {
+    router.get('/usuario/:id', verifyToken, async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
+      try {
+        const id = req.params.id;
+        const usuario = await Usuario.findById(id);
         if (!usuario) {
           res.status(404).send('Usuário não encontrado');
         } else {
           res.send(usuario);
         }
-      })
-      .catch((error) => {
+      }
+      catch(error) {
         res.status(500).send('Erro ao obter usuário: ' + error);
-      });
-  });
-    
-    router.post('/usuarios', (req, res) => {
-    const usuario = new Usuario(req.body);
-    usuario.save()
-      .then(() => {
-        res.status(201).send('Usuário criado com sucesso');
-      })
-      .catch((error) => {
-        res.status(400).send('Erro ao criar usuário: ' + error);
-      });
-  });
+      };
+    });
+
+    router.post('/usuario', (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
+      const usuario = new Usuario(req.body);
+      usuario.save()
+        .then(() => {
+          res.status(201).send('Usuário criado com sucesso');
+        })
+        .catch((error) => {
+          res.status(400).send('Erro ao criar usuário: ' + error);
+        });
+    });
   
   
-  router.put('/usuarios/:id', (req, res) => {
+  router.put('/usuario/:id', verifyToken, (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const id = req.params.id;
     Usuario.findByIdAndUpdate(id, req.body)
       .then(() => {
@@ -77,7 +89,13 @@ const options = {
   });
   
  
-  router.get('/usuarios', (req, res) => {
+  router.get('/usuario', verifyToken, async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
+    if(tokenChallenge(req.token, res))
+      return res.sendStatus(401);
+
     Usuario.find()
       .then((usuarios) => {
         res.send(usuarios);
@@ -87,8 +105,35 @@ const options = {
       });
   });
   
+  router.post('/usuario/login', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
+    try {
+      const { email, senha } = req.body;
+      const usuario = await Usuario.findOne({ email });
+      if (usuario) {
+        const result = bcrypt.compare(senha, usuario.senha);
+        if (result) {
+          const token = jwt.sign({ email: usuario.email }, jwtSECRET);
+          res.json({ token });
+        } 
+        else {
+          res.status(400).json({ error: "Senha inválida" });
+        }
+      } 
+      else {
+        res.status(400).json({ error: "Email inválido" });
+      }
+    } catch (error) {
+      res.status(400).json({ error });
+    }
+  })
 
-  router.delete('/usuarios/:id', (req, res) => {
+  router.delete('/usuario/:id', verifyToken, (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const id = req.params.id;
     Usuario.findByIdAndDelete(id)
       .then(() => {
@@ -106,6 +151,9 @@ const options = {
   //API UsuarioResponsavel
 
   router.post('/usuarios-responsaveis', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     try {
       const usuarioResponsavel = await UsuarioResponsavel.create(req.body);
       res.status(201).json(usuarioResponsavel);
@@ -116,6 +164,9 @@ const options = {
   
 
   router.get('/usuarios-responsaveis', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     try {
       const usuariosResponsaveis = await UsuarioResponsavel.find();
       res.json(usuariosResponsaveis);
@@ -126,6 +177,9 @@ const options = {
   
   
   router.get('/usuarios-responsaveis/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const usuarioResponsavel = await UsuarioResponsavel.findById(id);
@@ -141,6 +195,9 @@ const options = {
   
   
   router.put('/usuarios-responsaveis/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const usuarioResponsavel = await UsuarioResponsavel.findByIdAndUpdate(id, req.body, { new: true });
@@ -156,6 +213,9 @@ const options = {
   
   
   router.delete('/usuarios-responsaveis/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const usuarioResponsavel = await UsuarioResponsavel.findByIdAndDelete(id);
@@ -177,6 +237,9 @@ const options = {
 
 
   router.post('/contatos-usuario', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     try {
       const contatoUsuario = await ContatoUsuario.create(req.body);
       res.status(201).json(contatoUsuario);
@@ -187,6 +250,9 @@ const options = {
   
 
   router.get('/contatos-usuario', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     try {
       const contatosUsuario = await ContatoUsuario.find();
       res.json(contatosUsuario);
@@ -197,6 +263,9 @@ const options = {
   
  
   router.get('/contatos-usuario/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const contatoUsuario = await ContatoUsuario.findById(id);
@@ -212,6 +281,9 @@ const options = {
   
 
   router.put('/contatos-usuario/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const contatoUsuario = await ContatoUsuario.findByIdAndUpdate(id, req.body, { new: true });
@@ -227,6 +299,9 @@ const options = {
   
  
   router.delete('/contatos-usuario/:id', async (req, res) => {
+    if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
     const { id } = req.params;
     try {
       const contatoUsuario = await ContatoUsuario.findByIdAndDelete(id);
@@ -245,6 +320,9 @@ const options = {
     // API Figura
 
     router.get('/figuras', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const figuras = await Figura.find();
         res.json(figuras);
@@ -255,6 +333,9 @@ const options = {
     
     
     router.post('/figuras', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const novaFigura = new Figura(req.body);
         await novaFigura.save();
@@ -266,6 +347,9 @@ const options = {
     
  
     router.get('/figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const figura = await Figura.findById(req.params.id);
         if (figura) {
@@ -280,6 +364,9 @@ const options = {
     
    
     router.put('/figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const figura = await Figura.findByIdAndUpdate(req.params.id, req.body, { new: true });
         if (figura) {
@@ -294,6 +381,9 @@ const options = {
     
    
     router.delete('/figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const figura = await Figura.findByIdAndDelete(req.params.id);
         if (figura) {
@@ -311,6 +401,9 @@ const options = {
     //API Categoria
 
     router.get('/categoria', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categorias = await Categoria.find();
         res.json(categorias);
@@ -320,6 +413,9 @@ const options = {
     });
     
     router.get('/categoria/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoria = await Categoria.findById(req.params.id);
         if (categoria) {
@@ -333,6 +429,9 @@ const options = {
     });
     
     router.post('/categoria', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const novaCategoria = new Categoria(req.body);
         await novaCategoria.save();
@@ -343,6 +442,9 @@ const options = {
     });
     
     router.put('/categoria/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoria = await Categoria.findByIdAndUpdate(req.params.id, req.body, { new: true });
         if (categoria) {
@@ -356,6 +458,9 @@ const options = {
     });
     
     router.delete('/categoria/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoria = await Categoria.findByIdAndDelete(req.params.id);
         if (categoria) {
@@ -373,6 +478,9 @@ const options = {
     //API CategoriaFigura
 
     router.get('/categoria-figuras', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoriaFiguras = await CategoriaFigura.find();
         res.json(categoriaFiguras);
@@ -382,6 +490,9 @@ const options = {
     });
     
     router.post('/categoria-figuras', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const novaCategoriaFigura = new CategoriaFigura(req.body);
         await novaCategoriaFigura.save();
@@ -392,6 +503,9 @@ const options = {
     });
     
     router.get('/categoria-figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoriaFigura = await CategoriaFigura.findById(req.params.id);
         if (categoriaFigura) {
@@ -405,6 +519,9 @@ const options = {
     });
     
     router.put('/categoria-figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoriaFigura = await CategoriaFigura.findByIdAndUpdate(req.params.id, req.body, { new: true });
         if (categoriaFigura) {
@@ -418,6 +535,9 @@ const options = {
     });
     
     router.delete('/categoria-figuras/:id', async (req, res) => {
+      if(tokenChallenge(req.token, res))
+        return res.sendStatus(401);
+        
       try {
         const categoriaFigura = await CategoriaFigura.findByIdAndDelete(req.params.id);
         if (categoriaFigura) {
@@ -432,7 +552,26 @@ const options = {
 
     //Fim CategoriaFigura
 
+  function verifyToken(req,res,next){
+    const bearerHeader = req.headers['authorization'];
+    if(bearerHeader){
+        const bearer = bearerHeader.split(' ');
+        const bearerToken = bearer[1];
+        req.token = bearerToken;
 
-  app.use('/',router);
-  app.listen(3000)
-  console.log("Servidor funcionando ")
+        next();
+    }else{
+        res.sendStatus(401);
+    }
+  }
+
+  function tokenChallenge(token, res){
+    jwt.verify(token, 'secretKey', (err, authData) => {
+      if (err)
+        return !!err;
+    })
+  }
+
+app.use('/',router);
+app.listen(3000)
+console.log("Servidor funcionando")
